+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2146 - As admin I can check 'Participants must evaluate happening'
+
 ## [2.82.3] - 2021-03-05
 ### Fixed
 - MV-311 - fix max size upload video 300M