+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- MV-313 - trunck happening name if too long for tokbox
+
 ## [2.82.3] - 2021-03-05
 ### Fixed
 - MV-311 - fix max size upload video 300M